fixed server side rendering error
